fix layer button block
