+++ web/src/services/i18n/translations/en.yml
@@ -257,7 +257,7 @@ Spacing settings: Spacing settings
 New Camera: New Camera
 New Camera Button: New Camera Button
 Buttons List: Buttons List
-New Layer Button: New Layer Button
+New Layers Button: New Layers Button
 Add markdown text here: Add markdown text here
 Unsupported field type: Unsupported field type
 Drop here: Drop here
