+++ web/src/services/i18n/translations/ja.yml
@@ -236,7 +236,7 @@ Spacing settings: Spacing設定
 New Camera: 新しいカメラ
 New Camera Button: 新カメラボタン
 Buttons List: ボタン一覧
-New Layer Button: 新レイヤーボタン
+New Layers Button: 新レイヤーボタン
 Add markdown text here: ''
 Unsupported field type: ''
 Drop here: ここにドロップ
